Refactor: gather type generators in one file
